Move common utils to common module
