+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/util/FlowUtils.kt
@@ -1,27 +1,16 @@
 package net.mullvad.mullvadvpn.util
 
-import android.content.ComponentName
-import android.content.Context
-import android.content.Intent
-import android.content.ServiceConnection
-import android.os.IBinder
-import android.util.Log
 import android.view.animation.Animation
 import kotlin.coroutines.EmptyCoroutineContext
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.take
-import net.mullvad.mullvadvpn.model.ServiceResult
+import net.mullvad.mullvadvpn.lib.common.util.safeOffer
 import net.mullvad.mullvadvpn.ui.serviceconnection.ServiceConnectionState
 import net.mullvad.talpid.util.EventNotifier
-
-fun <T> SendChannel<T>.safeOffer(element: T): Boolean {
-    return runCatching { trySend(element).isSuccess }.getOrDefault(false)
-}
 
 fun Animation.transitionFinished(): Flow<Unit> =
     callbackFlow<Unit> {
@@ -40,34 +29,6 @@ fun Animation.transitionFinished(): Flow<Unit> =
             }
         }
         .take(1)
-
-fun Context.bindServiceFlow(intent: Intent, flags: Int = 0): Flow<ServiceResult> = callbackFlow {
-    val connectionCallback =
-        object : ServiceConnection {
-            override fun onServiceConnected(className: ComponentName, binder: IBinder) {
-                safeOffer(ServiceResult(binder))
-            }
-
-            override fun onServiceDisconnected(className: ComponentName) {
-                safeOffer(ServiceResult.NOT_CONNECTED)
-                bindService(intent, this, flags)
-            }
-        }
-
-    bindService(intent, connectionCallback, flags)
-
-    awaitClose {
-        safeOffer(ServiceResult.NOT_CONNECTED)
-
-        Dispatchers.Default.dispatch(EmptyCoroutineContext) {
-            try {
-                unbindService(connectionCallback)
-            } catch (e: IllegalArgumentException) {
-                Log.e("mullvad", "Cannot unbind as no binding exists.")
-            }
-        }
-    }
-}
 
 fun <R> Flow<ServiceConnectionState>.flatMapReadyConnectionOrDefault(
     default: Flow<R>,
